Rename fancy folder to select
